refactor: centralize integration test helpers
