+++ tests/integration/high_load_queue_test.go
@@ -21,7 +21,7 @@ func TestIntegrationHighLoadQueue(testingInstance *testing.T) {
 	router, buildRouterError := proxy.BuildRouter(proxy.Configuration{
 		ServiceSecret: serviceSecretValue,
 		OpenAIKey:     openAIKeyValue,
-		LogLevel:      "debug",
+		LogLevel:      logLevelDebug,
 		WorkerCount:   1,
 		QueueSize:     proxy.DefaultQueueSize,
 	}, newLogger(testingInstance))
