+++ tests/integration/long_request_test.go
@@ -14,15 +14,8 @@ import (
 )
 
 const (
-	serviceSecretValue           = "sekret"
-	openAIKeyValue               = "sk-test"
-	mockModelsURL                = "https://mock.local/v1/models"
-	mockResponsesURL             = "https://mock.local/v1/responses"
 	modelsListBody               = `{"data":[{"id":"gpt-4.1"}]}`
 	expectedResponseBody         = "SLOW_OK"
-	promptQueryParameter         = "prompt"
-	keyQueryParameter            = "key"
-	promptValue                  = "ping"
 	responseDelay                = 31 * time.Second
 	httpClientTimeout            = responseDelay + 5*time.Second
 	requestTimeoutSecondsDefault = 40
@@ -55,7 +48,7 @@ func TestIntegrationResponseDeliveredAfterDelay(testingInstance *testing.T) {
 	for _, testCase := range testCases {
 		testingInstance.Run(testCase.name, func(subTest *testing.T) {
 			configureProxy(subTest, makeSlowHTTPClient(subTest))
-			router, buildError := proxy.BuildRouter(proxy.Configuration{ServiceSecret: serviceSecretValue, OpenAIKey: openAIKeyValue, LogLevel: "debug", WorkerCount: 1, QueueSize: 8, RequestTimeoutSeconds: requestTimeoutSecondsDefault}, newLogger(subTest))
+			router, buildError := proxy.BuildRouter(proxy.Configuration{ServiceSecret: serviceSecretValue, OpenAIKey: openAIKeyValue, LogLevel: logLevelDebug, WorkerCount: 1, QueueSize: 8, RequestTimeoutSeconds: requestTimeoutSecondsDefault}, newLogger(subTest))
 			if buildError != nil {
 				subTest.Fatalf("BuildRouter failed: %v", buildError)
 			}
